% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ilogit}
\alias{ilogit}
\title{The inverse logit function}
\usage{
ilogit(log_odds)
}
\arguments{
\item{log_odds}{The logarithm, to base e, of an odds}
}
\value{
A probability, i.e. the p in log(p/(1-p))
}
\description{
This function calculates the inverse of the log-odds.
The log odds of a probability p is log_odds = log(p/(1-p)), so
ilogit(log_odds) = p. It is already built in to R with the `plogis`
function, but I prefer to code it explicitly and call it `ilogit`.
}
\examples{
p <- 0.75
log_odds <- log(p/(1-p))
q <- ilogit(log_odds)
p == q # Should be TRUE
}
\author{
Mark Andrews (mark.andrews@ntu.ac.uk)
}
